feat(login): boilerplate migration & dockerisation
