+++ .env.example
@@ -17,7 +17,7 @@ PHP_CLI_SERVER_WORKERS=4
 BCRYPT_ROUNDS=12
 
 LOG_CHANNEL=stack
-LOG_STACK=single
+LOG_STACK=daily
 LOG_DEPRECATIONS_CHANNEL=null
 LOG_LEVEL=debug
 
@@ -57,10 +57,8 @@ MAIL_PASSWORD=null
 MAIL_FROM_ADDRESS="hello@example.com"
 MAIL_FROM_NAME="${APP_NAME}"
 
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_DEFAULT_REGION=us-east-1
-AWS_BUCKET=
-AWS_USE_PATH_STYLE_ENDPOINT=false
-
 VITE_APP_NAME="${APP_NAME}"
+
+TELESCOPE_ENABLED=false
+TELESCOPE_DOMAIN="${APP_URL}"
+TELESCOPE_PATH=telescope
