+++ app/Livewire/Admin/Layouts/Navigation.php
@@ -2,6 +2,18 @@
 
 namespace App\Livewire\Admin\Layouts;
 
+use App\Livewire\Admin\Actions\Logout;
 use Livewire\Component;
 
-class Navigation extends Component {}
+class Navigation extends Component
+{
+    /**
+     * Log the current user out of the application.
+     */
+    public function logout(Logout $logout): void
+    {
+        $logout();
+
+        $this->redirect('/', navigate: true);
+    }
+}
