+++ app/Livewire/Admin/Pages/Index.php
@@ -2,13 +2,14 @@
 
 namespace App\Livewire\Admin\Pages;
 
+use Illuminate\Contracts\View\View;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 
 #[Layout('layouts.admin')]
 class Index extends Component
 {
-    public function render()
+    public function render(): View
     {
         return view('livewire.admin.pages.index');
     }
