+++ app/Livewire/Admin/Pages/Profile/DeleteUserForm.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Livewire\Admin\Pages\Profile;
+
+use App\Livewire\Admin\Actions\Logout;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Livewire\Component;
+
+class DeleteUserForm extends Component
+{
+    public string $password = '';
+
+    /**
+     * Delete the currently authenticated user.
+     */
+    public function deleteUser(Logout $logout): void
+    {
+        $this->validate([
+            'password' => ['required', 'string', 'current_password'],
+        ]);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        tap($user, $logout(...))->delete();
+
+        $this->redirect('/', navigate: true);
+    }
+}
