+++ app/Livewire/Admin/Pages/Profile/Profile.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Livewire\Admin\Pages\Profile;
+
+use Livewire\Attributes\Layout;
+use Livewire\Component;
+
+#[Layout('layouts.admin')]
+class Profile extends Component {}
