+++ app/Livewire/Admin/Pages/Profile/UpdatePasswordForm.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Livewire\Admin\Profile;
+namespace App\Livewire\Admin\Pages\Profile;
 
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules\Password;
@@ -32,7 +33,10 @@ class UpdatePasswordForm extends Component
             throw $e;
         }
 
-        Auth::user()->update([
+        /** @var User $user */
+        $user = Auth::user();
+
+        $user->update([
             'password' => Hash::make($validated['password']),
         ]);
