+++ app/Livewire/Blog/Pages/Index.php
@@ -2,11 +2,12 @@
 
 namespace App\Livewire\Blog\Pages;
 
+use Illuminate\Contracts\View\View;
 use Livewire\Component;
 
 class Index extends Component
 {
-    public function render()
+    public function render(): View
     {
         return view('livewire.blog.pages.index');
     }
