+++ app/View/Components/LoginLayout.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\View\Components;
+
+use Illuminate\View\Component;
+
+class LoginLayout extends Component
+{
+    /**
+     * Get the view / contents that represents the component.
+     */
+    public function render()
+    {
+        return view('layouts.login');
+    }
+}
