+++ composer.json
@@ -16,11 +16,13 @@
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^3.5",
+        "brianium/paratest": "^7.7",
         "fakerphp/faker": "^1.23",
         "larastan/larastan": "3.0",
         "laravel/breeze": "^2.3",
         "laravel/pail": "^1.1",
         "laravel/pint": "^1.13",
+        "laravel/sail": "^1.41",
         "laravel/telescope": "^5.4",
         "mockery/mockery": "^1.6",
         "nunomaduro/collision": "^8.1",
