+++ config/database.php
@@ -15,7 +15,6 @@ return [
     | is explicitly specified when you execute a query / statement.
     |
     */
-
     'default' => env('DB_CONNECTION', 'sqlite'),
 
     /*
