+++ resources/views/livewire/admin/pages/profile/profile.blade.php
@@ -1,0 +1,27 @@
+<x-slot name="header">
+    <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">
+        {{ __('Profile') }}
+    </h2>
+</x-slot>
+
+<div class="py-12">
+    <div class="max-w-7xl mx-auto sm:px-6 lg:px-8 space-y-6">
+        <div class="p-4 sm:p-8 bg-white dark:bg-gray-800 shadow sm:rounded-lg">
+            <div class="max-w-xl">
+                <livewire:admin.pages.profile.update-profile-information-form/>
+            </div>
+        </div>
+
+        <div class="p-4 sm:p-8 bg-white dark:bg-gray-800 shadow sm:rounded-lg">
+            <div class="max-w-xl">
+                <livewire:admin.pages.profile.update-password-form/>
+            </div>
+        </div>
+
+        <div class="p-4 sm:p-8 bg-white dark:bg-gray-800 shadow sm:rounded-lg">
+            <div class="max-w-xl">
+                <livewire:admin.pages.profile.delete-user-form/>
+            </div>
+        </div>
+    </div>
+</div>
