--- tests/Feature/Auth/RegistrationTest.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Tests\Feature\Auth;
-
-use Livewire\Volt\Volt;
-
-test('registration screen can be rendered', function () {
-    $response = $this->get('/register');
-
-    $response
-        ->assertOk()
-        ->assertSeeVolt('pages.auth.register');
-});
-
-test('new users can register', function () {
-    $component = Volt::test('pages.auth.register')
-        ->set('name', 'Test User')
-        ->set('email', 'test@example.com')
-        ->set('password', 'password')
-        ->set('password_confirmation', 'password');
-
-    $component->call('register');
-
-    $component->assertRedirect(route('dashboard', absolute: false));
-
-    $this->assertAuthenticated();
-});
